\name{computeFromRuler}
\alias{computeFromRuler}
\title{compute the new index directly according to the ruler parameter}
\usage{
dt.index <- computeFromRuler(dt, id_colnum, dt.para, category=NA)
}
\description{
This function is used to compute the new index using ruler information (i.e. dt.para parameter). The category parameter need to be defined when the ruler contains category.
}
\examples{
NA
}
