\name{getDiscriminationIndex}
\alias{getDiscriminationIndex}
\title{get the discrimination index from the index data}
\usage{
remark <- getDiscriminationIndex(dt.index, index_colname)
}
\description{
This function is used to compute the discrimination index which evluates the multi-index rating system.
}
\examples{
NA
}
