\name{getIndexSystem}
\alias{getIndexSystem}
\title{get the whole index system based on the known and standlized index data}
\usage{
dt.index <- getIndexSystem(dt, weight, NAnum=10)
}
\description{
This function is used to automatically compute the higher index using basic index data according to the weighted.mean method.
The function allow to define your NA item handling stratigy by defining the max allowed NA number: NAnum.
}
\examples{
NA
}
