\name{getRulerPara}
\alias{getRulerPara}
\title{get Ruler Parameters for further computing}
\usage{
dt.para <- getRulerPara(dt, index_colrange, category_colnum=NA)
}
\description{
This function is used to compute the ruler parameters which can help to quickly compute the index of new data using computeFromRuler function.
Note that the parameter category_colnum is used to define the ruler according to some category.
}
\examples{
NA
}
