\name{getStandlized}
\alias{getStandlized}
\title{make the index value range standlize to [0,1]}
\usage{
dt.index <- getStandlized(dt.index, index_colrange, id_colnum, category_colnum=NA)
}
\description{
This function is used to standlize any index values to [0,1].
The function also need to assign the id column number, index column range, and category column number if you want to ranking the index according to it.
}
\examples{
NA
}
