\name{indexFunc}
\alias{indexFunc}
\title{compute some basic indexes automatically}
\usage{
1. the index system should have columns: func and parameter
2. prepare the data to specific format: <id, date, value>
3. use getFunction() to load the needed functions and then compute
}
\description{
These functions are helped to compute some basic indexes automatically, including:
Note: the default granularity is month
- portion index
- yoy index
- qoq index
- modified yoy index (yoyplus)
- modified qoq index (qoqplus)
- monthVol index
- seasonVol index
- yearVol index
- monthHhi index
- to be added
}
\examples{
NA
}
